02_Projects/obsidian_rag/requirements.txt
chromadb>=0.5.0
google-genai>=1.0.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
watchdog>=4.0.0
tiktoken>=0.7.0
pyyaml>=6.0.1
python-multipart>=0.0.9
